requirements.txt
asyncio
discord.py==1.3.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
emoji==0.5.4
fuzzywuzzy==0.18.0